// src/app/@core/model/rest/company-admin/user-management/users-list-res.ts
export interface UsersListRes {
  companyUserList: CompanyUserList[]
}

export interface CompanyUserList {
  userId: string
  blocked: string
  firstLogin: boolean
  tries: number
  maxTries: number
  questionTries: any
  maxQuestionTries: any
  userName: string
  userNameArabic: any
  nickname: any
  status: any
  type: string
  idIqama: string
  department: any
  address: any
  phone: any
  mobile: string
  secondMobile: any
  email: string
  secondEmail: any
  nationality: any
  region: string
  city: string
  expiryDate: any
  issuingPlace: any
  active: any
  lastLogon: any
  lastPasswordChange: string
  userImage: any
  authenticationMethod: string
  passport: any
  preferedLanguage: any
  createdBy: string
  idExpireDate: any
  createdDate: string
  birthDate: any
  passwordExpired: boolean
  tutorialDone: boolean
  socialTwitter: any
  socialFacebook: any
  socialInstagram: any
  socialLinkedin: any
  userStatus: string
  qtlLimit: number
  companyQtlLimit: any
  showSMSAlert: any
  companyPk: number
  userPk: number
  empRef: string
  ownacclimit: any
  withinlimit: any
  locallimit: any
  internationallimit: any
  lastLoginDate: any
  lastLoginTime: any
  limit: any
  limitSpent: number
  dateLimitReset: any
  accessLimited: boolean
  accessFrom: any
  accessTo: any
  billPaymentLimit: any
  sadadInvoiceHubLimit: any
  governmentPaymentLimit: any
  billLimitSpent: number
  sadadInvoiceHubLimitSpent: number
  governmentLimitSpent: number
  billDateLimitReset: any
  sadadInvoiceHubDateLimitReset: any
  governmentDateLimitReset: any
  alertsFeeAccount: any
  fundLimit: any
  fundLimitSpent: number
  fundDateLimitReset: any
  bulkLimit: any
  bulkLimitSpent: number
  bulkDateLimitReset: any
  tokenSerial: any
  tokenStatus: any
  tokenSuccessLogin: any
  tokenFailLogin: any
  loginMethodType: any
  tokenLanguage: any
  passDelivery: any
  userAuthenticationType: any
  softTokenSubstatus: any
  groups: string[]
  daysOfWeek: any[]
  cardDepartments: any
  vaPermissions: any
  tokenDevice: any
  process: any
  processStatus: any
  urPayLimit: any
  maker?: boolean;
  checker?: boolean;
}