// src/app/@core/model/rest/company-admin/user-management/users-list-req.ts
export interface UsersListReq {
    department?: any
    mobileNumber?: any
    userId?: any
    city?: any
    region?: any
    userType?: any
    iqama?: any
    nickname?: any
    userName?: any
    empRef?: any
    status?: any
    createdBy?: any
    creationDateFrom?: any
    creationDateTo: string
    orderType: string
    lessDetails?: boolean;
  }